Add files via upload
This file has the source code of the project in which the dataset has been used is Flickr 8k dataset which is easily available on the internet. Also you will need to download the weights of VGG16 model and make a file which will store the feature vectors of the images.